requirements.txt
Django==3.1
django-admin-sortable2==0.7.7
Pillow==7.2.0
